requirements.txt
Django==4.2.1

djangorestframework~=3.14.0
pip~=21.1.2
attrs~=23.1.0
wheel~=0.36.2
setuptools~=57.0.0